feat(arcan): unify liquid prompt across daemon and shell (BRO-366, BRO-374, BRO-375)
Move the liquid prompt builder from the `arcan` binary crate to
`arcan-core` so both the shell REPL and the daemon HTTP server share
the same prompt assembly logic.

- Move `prompt.rs` to `arcan-core::prompt` (public module)
- Make key functions public: `build_git_section`, `build_environment_section`,
  `build_memory_section`, `load_project_instructions`, etc.
- Add `chrono` + `tempfile` (dev) to arcan-core dependencies
- Replace `arcan/src/prompt.rs` with thin re-export of `arcan_core::prompt`
- Add `workspace_root` parameter to `create_canonical_router_with_skills`
- Daemon now injects: environment info, git context (per-request), project
  instructions (cached at startup), memory, and guidelines into every tick
- Add `test_prompt_available_from_core` verifying the public API surface
- All 99 arcan-core tests pass, all workspace tests pass, zero clippy warnings
